requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
imbalanced-learn==0.12.4
xgboost==2.1.1
kafka-python==2.0.2
joblib==1.4.2
PyMySQL==1.1.1
cryptography==46.0.3
python-dotenv==1.0.1
